requirements.txt
# Core dependencies
python>=3.8
torch>=1.13.1
transformers>=4.25.0
datasets>=2.8.0
wandb>=0.13.0
sentencepiece>=0.1.97
Biopython>=1.79
numpy>=1.22.0
tqdm>=4.64.0
pyyaml>=6.0

# Mamba specific dependencies
mamba-ssm>=1.0.0

# Optional dependencies
matplotlib>=3.5.0
scipy>=1.7.0
pandas>=1.3.0